move checksum and tag file writing to BagItArchive.cs
