+++ lib/FiveSafes.Net/FiveSafes.Net/BagItArchive.cs
@@ -1,7 +1,13 @@
+using FiveSafes.Net.Constants;
+using FiveSafes.Net.Utilities;
+
 namespace FiveSafes.Net;
 
 public class BagItArchive
 {
+  private static string[] _tagFiles =
+    { BagItConstants.BagitTxtPath, BagItConstants.BagInfoTxtPath, BagItConstants.ManifestPath };
+
   private DirectoryInfo _archive;
   private string _payloadDirectoryPath = "data";
 
@@ -40,5 +46,72 @@ public class BagItArchive
     if (!sourceFileInfo.Exists) throw new FileNotFoundException();
     var destination = Path.Combine(toPayload ? PayloadDirectoryPath : ArchiveRootPath, sourceFileInfo.Name);
     sourceFileInfo.CopyTo(destination, overwrite: true);
+  }
+
+  /// <summary>
+  /// Compute the SHA512 for each file in the Bagit archive's <c>data</c> subdirectory and write a
+  /// <c>manifest-sha512.txt</c> to the archive.
+  /// </summary>
+  public async Task WriteManifestSha512()
+  {
+    await using var manifestFile =
+      new FileStream(Path.Combine(ArchiveRootPath, BagItConstants.ManifestPath), FileMode.Create,
+        FileAccess.Write);
+    if (!Directory.Exists(PayloadDirectoryPath)) return;
+    await using var writer = new StreamWriter(manifestFile);
+    foreach (var entry in Directory.EnumerateFiles(PayloadDirectoryPath, "*", SearchOption.AllDirectories))
+    {
+      await using var stream = new FileStream(entry, FileMode.Open, FileAccess.Read);
+      var checksum = ChecksumUtility.ComputeSha512(stream);
+      // Note there should be 2 spaces between the checksum and the file path
+      // The path should be relative to bagitDir
+      var path = Path.GetRelativePath(ArchiveRootPath, entry);
+      await writer.WriteLineAsync($"{checksum}  {path}");
+    }
+  }
+
+  /// <summary>
+  /// Compute the SHA512 for the <c>bagit.txt</c>, <c>bag-info.txt</c> and <c>manifest-sha512.txt</c> and
+  /// write a <c>tagmanifest-sha512.txt</c> to the archive.
+  /// </summary>
+  public async Task WriteTagManifestSha512()
+  {
+    await using var manifestFile =
+      new FileStream(Path.Combine(ArchiveRootPath, BagItConstants.TagManifestPath), FileMode.Create,
+        FileAccess.Write);
+    await using var writer = new StreamWriter(manifestFile);
+    foreach (var tagFile in _tagFiles)
+    {
+      var filePath = Path.Combine(ArchiveRootPath, tagFile);
+      if (!File.Exists(filePath)) continue;
+      await using var stream = new FileStream(filePath, FileMode.Open, FileAccess.Read);
+      var checksum = ChecksumUtility.ComputeSha512(stream);
+      // Note there should be 2 spaces between the checksum and the file path
+      await writer.WriteLineAsync($"{checksum}  {tagFile}");
+    }
+  }
+
+  public async Task WriteBagitTxt()
+  {
+    string[] contents = { "BagIt-Version: 1.0", "Tag-File-Character-Encoding: UTF-8" };
+    await using var manifestFile =
+      new FileStream(Path.Combine(ArchiveRootPath, BagItConstants.BagitTxtPath), FileMode.Create,
+        FileAccess.Write);
+    await using var writer = new StreamWriter(manifestFile);
+    foreach (var line in contents)
+    {
+      await writer.WriteLineAsync(line);
+    }
+  }
+
+  public async Task WriteBagInfoTxt()
+  {
+    var contents = "External-Identifier: urn:uuid:{0}";
+    var line = string.Format(contents, Guid.NewGuid().ToString());
+    await using var manifestFile =
+      new FileStream(Path.Combine(ArchiveRootPath, BagItConstants.BagInfoTxtPath), FileMode.Create,
+        FileAccess.Write);
+    await using var writer = new StreamWriter(manifestFile);
+    await writer.WriteLineAsync(line);
   }
 }
